+++ src/main.rs
@@ -12,39 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-extern crate kytan;
-
-use kytan::common::utils;
-use kytan::network::network;
-use kytan::cmd::cli;
-
-
-use std::sync::atomic::Ordering;
-use env_logger;
-use libc;
-
-
-extern "C" fn handle_signal(_: libc::c_int) {
-    network::INTERRUPTED.store(true, Ordering::Relaxed);
-}
+use kytan::run_app;
 
 fn main() {
-    env_logger::init();
-
-    if !utils::is_root() {
-        panic!("Please run as root");
-    }
-
-
-    unsafe {
-        libc::signal(libc::SIGINT, handle_signal as libc::sighandler_t);
-        libc::signal(libc::SIGTERM, handle_signal as libc::sighandler_t);
-    }
-
-    match cli::get_args().unwrap() {
-        cli::Args::Client(client) => network::connect(&client.remote_addr, client.port, client.default_route, &client.key),
-        cli::Args::Server(server) => network::serve(server.port, &server.key, server.dns),
-    }
-
-    println!("SIGINT/SIGTERM captured. Exit.");
+    run_app();
 }
